fix:file path update and settings update
