+++ .gitignore
@@ -11,5 +11,5 @@ main.exe
 main
 
 static/uploadfile/*
-temp/logs/*
+logs/*
 tmp/*
